# backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Betel Hospitality API",
    "description": "Comprehensive API documentation for Betel Hospitality Platform (Hotels, Tours, Cruises, Buses)",
    "version": "1.0.0",
    "contact": {
      "name": "Dev Team",
      "email": "dev@betelhospitality.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    },
    {
      "url": "https://localhost:8080/"
    }
  ],
...
  "paths": {
...
    "/api/hotels/hotel/{id}": {
      "delete": {
        "tags": [
          "Hotels"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/hotels/hotel": {
      "patch": {
        "tags": [
          "Hotels"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "_id": {
                    "example": "any"
                  },
                  "roomType": {
                    "example": "any"
                  },
                  "cheapestPrice": {
                    "example": "any"
                  },
                  "services": {
                    "example": "any"
                  },
                  "stars": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/hotels/{slug}": {
      "get": {
        "tags": [
          "Hotels"
        ],
        "description": "",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/cars/route": {
      "post": {
        "tags": [
          "car"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "origin": {
                    "example": "any"
                  },
                  "destination": {
                    "example": "any"
                  },
                  "prices": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/cars/route/{id}": {
      "get": {
        "tags": [
          "car"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "car"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "car"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/cars/routes": {
      "get": {
        "tags": [
          "car"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}